Move new model to fault-marking shared model
